Split updater in several files
- Updater is now splitted in several files,
  each for every source
- use ' instead of " in all strings
